requirements.txt
python-dotenv
watchdog==5.0.2
pyicloud=1.0.0

